Refactor SystemService and ComposerProcessService class.
